Consolidate project_db modules: schema_gen→schema, coerce+case_adapter→populate
Merge schema_gen into schema and coerce+case_adapter into populate to
reduce module count. Organize each file by the newspaper metaphor:
public API at top, private helpers below. No logic changes.
